refactor: move api into ~lib directory
