+++ frontend/src/features/auth/api/logout.ts
@@ -1,4 +1,4 @@
-import { api } from '~/api'
+import { api } from '~lib/api'
 
 export const logout = async (): Promise<void> => {
   return api.url('/auth/logout').post().json()
