+++ frontend/src/features/auth/api/verifyLoginOtp.ts
@@ -3,7 +3,7 @@ import {
   VerifyOtpResponseDto,
 } from '~shared/types/auth.dto'
 
-import { api } from '~/api'
+import { api } from '~lib/api'
 
 export const verifyLoginOtp = async (params: VerifyOtpRequestDto) => {
   return api.url('/auth/verify').post(params).json<VerifyOtpResponseDto>()
